Discard GPS data received during sleep cycle
Currently, when serial data is received during the sleep cycle, the cpu
wakes, but does not read it.  As a result, this data piles up in the
serial buffer, or is discarded once the buffer is full.

When the sleep cycle is complete, the cpu wakes and begins processing
the stale and truncated serial data in the buffer.

I haven't found this to cause an error, since the parser recognizes the
data as invalid.  But since we know the data is old/bad, why not just
get rid of it.

This patch simply discards any serial data received during the sleep
window.

I also considered flushing the buffer at the start of the wake cycle, but
this might discard the first character of a new message, if it was the
message that woke the cpu.