Small parsing cleanup
NMEA sentences end with a carriage return (\r) **and** linefeed (\n).
Currently, the parser recognizes the end of a sentence by looking for
**either** of these characters, so it thinks there are two
sentences where there is only one.

In practice, this does not cause a problem.  It is more of an annoyance
when developing, as both characters trigger the same code path which may
be instrumented with debugging.

This patch makes the parsing skip the \r, as it is redundant.