Initialize the parser at the start of a wake cycle
An earlier commit flushed the serial buffer at the start of the wake
cycle.  This patch adds to that by also resetting the state of the
parser.

Only an issue when processing messages, and you hit the
VALID_POS_TIMEOUT, and parsing stops in mid-message.  If the
parser is not reset at the beginning of the next wake cycle, the next
message will always be marked BAD.